Rename tab to tab-pane and tabs to tab
