+++ src/modules/Tab/index.js
@@ -1,2 +1,2 @@
+export { default as TabPane } from './TabPane';
 export { default as Tab } from './Tab';
-export { default as Tabs } from './Tabs';
